project/requirements.txt
fastapi
uvicorn[standard]
websockets
python-multipart
pydantic
pyyaml
python-dotenv
mcp
requests